Letting Convert to Base 64 more dinamic
