+++ Application/Vitorio.CLI/Commands/Convert/ConvertCommand.cs
@@ -5,7 +5,7 @@ public class ConvertCommand : ICommandFactory
     public Command Create()
     {
         Command command = new("convert", "Converte um dado de um tipo para outro");
-        command.AddCommand(new ConvertImageToBae64Command().Create());
+        command.AddCommand(new ConvertToBae64Command().Create());
         return command;
     }
 }
